+++ mysqldb.py
@@ -2,8 +2,8 @@ import mysql.connector
 def connector_mysql():
         mydb = mysql.connector.connect(
         host="localhost",
-        user="root",
-        password="vicky1598",
+        user="ilearnplace",
+        password="ilearnplace",
         database="ilearnplace"
         )
